Allow invocation without `GIT_DIR` variable
There are cases (`git ls-remote`) where no files are written to disk,
nor are any files read from disk. When Git invokes the remote helper
in these scenarios, `GIT_DIR` may be provided if the command is run
inside a worktree. However, this is not guaranteed - if the command
is run with no worktree then `GIT_DIR` remains unset and resolves to
an empty string in Go. The `os.MkdirAll` call should not happen
internally in such scenario.

This will enable Ansible to create a `Gitopia://` repository from
scratch, with no human interaction, as it does such checks internally.

Closes Gitopia/git-remote-gitopia#9.